requirements.txt
boto3==1.37.3
joblib==1.3.2
pandas==2.1.4
streamlit==1.42.2
python-dotenv==1.0.1
openai==1.65.1
instructor==1.7.2
langfuse==2.59.6
pydantic==2.10.6